split PyTorchSparkEstimator from estimator
